Move cmd package into commands
