--- cmd/cluster/scale/request/cluster.go
@@ -1,5 +1,0 @@
-package request
-
-type Cluster struct {
-	Scaling Scaling
-}
