--- cmd/cluster/scale/request/request.go
@@ -1,5 +1,0 @@
-package request
-
-type Request struct {
-	Cluster Cluster
-}
