--- cmd/cluster/scale/request/scaling.go
@@ -1,6 +1,0 @@
-package request
-
-type Scaling struct {
-	Max int64
-	Min int64
-}
