+++ commands/scale/cluster/command.go
@@ -12,9 +12,9 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/giantswarm/gsctl/client"
-	"github.com/giantswarm/gsctl/cmd/cluster/scale/defaulting"
-	"github.com/giantswarm/gsctl/cmd/cluster/scale/request"
 	"github.com/giantswarm/gsctl/commands/errors"
+	"github.com/giantswarm/gsctl/commands/scale/cluster/defaulting"
+	"github.com/giantswarm/gsctl/commands/scale/cluster/request"
 	"github.com/giantswarm/gsctl/config"
 	"github.com/giantswarm/gsctl/confirm"
 	"github.com/giantswarm/gsctl/flags"
